refactor to component
